780 move unused contracts and tests into unmaintained folder
